Revert "Productionize Audit Bucket Function (#1495)"
This reverts commit b44eec0058ca06bd967ac0a59bd6bd0909c13c6f.
