+++ .gitignore
@@ -55,5 +55,3 @@ services/app-api/lambda-layers-prisma-client-engine
 services/app-api/healthPlanFormDataMigrations
 
 services/app-proto/build
-
-testDefinitions/
