+++ dev_tool/src/local/graphql.ts
@@ -3,12 +3,20 @@ import { once } from '../deps.js'
 
 // run the graphql compiler with --watch
 async function compileGraphQLTypesWatch(runner: LabeledProcessRunner) {
+    await runner.runCommandAndOutput('gql deps', ['yarn', 'install'], '')
+
     runner.runCommandAndOutput('gqlgen', ['lerna', 'run', 'gqlgen:watch'], '')
 }
 
 export const compileGraphQLTypesWatchOnce = once(compileGraphQLTypesWatch)
 
 async function compileGraphQLTypes(runner: LabeledProcessRunner) {
+    await runner.runCommandAndOutput(
+        'gql deps',
+        ['yarn', 'install', '--prefer-offline'],
+        ''
+    )
+
     return runner.runCommandAndOutput('gqlgen', ['lerna', 'run', 'gqlgen'], '')
 }
