+++ dev_tool/src/local/s3.ts
@@ -1,10 +1,12 @@
 import LabeledProcessRunner from '../runner.js'
 
 // runS3Locally runs s3 locally
-export async function runUploadsLocally(runner: LabeledProcessRunner) {
+export async function runS3Locally(runner: LabeledProcessRunner) {
+    await runner.runCommandAndOutput('s3 yarn', ['yarn', 'install'], '')
+
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     runner.runCommandAndOutput(
-        'uploads',
+        's3',
         ['lerna', 'run', 'start', '--scope=uploads'],
         ''
     )
