+++ dev_tool/src/local/storybook.ts
@@ -1,8 +1,11 @@
 import LabeledProcessRunner from '../runner.js'
 import { compileGraphQLTypesWatchOnce } from './graphql.js'
+import { installWebDepsOnce } from './web.js'
 
 export async function runStorybookLocally(runner: LabeledProcessRunner) {
     compileGraphQLTypesWatchOnce(runner)
+
+    await installWebDepsOnce(runner)
 
     runner.runCommandAndOutput('storybook', ['lerna', 'run', 'storybook'], '')
 }
