+++ dev_tool/src/test/web.ts
@@ -4,6 +4,7 @@ import { spawn } from 'child_process'
 import {
     compileGraphQLTypesWatchOnce,
     compileGraphQLTypesOnce,
+    installWebDepsOnce,
     compileProto,
     compileProtoWatch,
 } from '../local/index.js'
@@ -13,6 +14,8 @@ export async function runWebTestsWatch(jestArgs: string[]) {
 
     await compileGraphQLTypesWatchOnce(runner)
     await compileProtoWatch(runner)
+
+    await installWebDepsOnce(runner)
 
     // because we are inheriting stdio for this process,
     // we need to not run spawnSync or else all the output
@@ -32,6 +35,7 @@ export async function runWebTests(
 ): Promise<number> {
     await compileGraphQLTypesOnce(runner)
     await compileProto(runner)
+    await installWebDepsOnce(runner)
 
     return await runner.runCommandAndOutput(
         'web - unit',
