+++ services/uploads/jest.config.ts
@@ -1,8 +1,6 @@
 module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
-    testTimeout: 60000, // Most of these test run clamscan which takes a while
-    globalSetup: '<rootDir>/src/jestGlobalSetup.ts',
     coverageReporters: [
         [
             'json',
@@ -20,5 +18,4 @@ module.exports = {
     ],
     moduleFileExtensions: ['js', 'json', 'jsx', 'd.ts', 'ts', 'node'],
     coveragePathIgnorePatterns: [],
-    modulePathIgnorePatterns: ['local_buckets'],
 }
