+++ services/uploads/serverless.yml
@@ -43,11 +43,6 @@ provider:
             - !Sub arn:aws:s3:::${self:service}-${sls:stage}-uploads-${AWS::AccountId}
             - !Sub arn:aws:s3:::${self:service}-${sls:stage}-avscan-${AWS::AccountId}
             - !Sub arn:aws:s3:::${self:service}-${sls:stage}-qa-${AWS::AccountId}
-        - Effect: 'Allow'
-          Action:
-            - lambda:InvokeFunction
-          Resource:
-            - !Sub arn:aws:lambda:${self:provider.region}:${AWS::AccountId}:function:${self:service}-${sls:stage}-avAuditFiles
 
 custom:
   region: ${aws:region}
@@ -67,7 +62,6 @@ custom:
     ssm:
       '/configuration/iam/full_permissions_boundary_policy': 'arn:aws:iam::local:policy/local/developer-boundary-policy'
   s3:
-    # address: 127.0.0.1 # this started defaulting to ipv6 which prevented tests from connecting
     # when vhostBuckets are enabled all non-localhost hostnames are re-written as buckets
     # setting this false allows the workaround for docker to work.
     vhostBuckets: false
@@ -79,7 +73,6 @@ custom:
         # the attachments bucket has no name, so we need to make one up on our own
       - local-qa
       - test-uploads
-      - test-audit
       - test-av-definitions
   serverlessTerminationProtection:
     stages:
@@ -95,7 +88,7 @@ layers:
 
 functions:
   avScan:
-    handler: src/lambdas/index.avScan
+    handler: src/avScan.avScanLambda
     name: ${self:service}-${sls:stage}-avScan
     timeout: 300 # 300 seconds = 5 minutes. Average scan is 25 seconds.
     memorySize: 3008
@@ -109,7 +102,7 @@ functions:
       OPENTELEMETRY_COLLECTOR_CONFIG_FILE: /var/task/collector.yml
 
   avDownloadDefinitions:
-    handler: src/lambdas/index.avDownloadDefinitions
+    handler: src/updateAVDefinitions.updateAVDefinitionsLambda
     events:
       - schedule: cron(0 */6 * * ? *)
     timeout: 300 # 300 seconds = 5 minutes
@@ -124,24 +117,8 @@ functions:
       OPENTELEMETRY_COLLECTOR_CONFIG_FILE: /var/task/collector.yml
 
   avAuditUploads:
-    handler: src/lambdas/index.avAuditUploads
-    timeout: 300 # 300 seconds = 5 minutes. Average scan is 25 seconds.
-    maximumRetryAttempts: 0
-    layers:
-      - !Ref ClamDefsLambdaLayer
-      - arn:aws:lambda:us-east-1:901920570463:layer:aws-otel-nodejs-amd64-ver-1-8-0:2
-    environment:
-      AUDIT_BUCKET_NAME: !Ref DocumentUploadsBucket
-      CLAMAV_BUCKET_NAME: !Ref ClamDefsBucket
-      PATH_TO_AV_DEFINITIONS: 'lambda/s3-antivirus/av-definitions'
-      LIST_INFECTED_FILES_LAMBDA_NAME: ${self:functions.avAuditFiles.name}
-      AWS_LAMBDA_EXEC_WRAPPER: /opt/otel-handler
-      OPENTELEMETRY_COLLECTOR_CONFIG_FILE: /var/task/collector.yml
-
-  avAuditFiles:
-    handler: src/lambdas/index.avAuditFiles
-    name: ${self:service}-${sls:stage}-avAuditFiles
-    timeout: 300 # 300 seconds = 5 minutes. Average scan is 25 seconds.
+    handler: src/auditUploads.auditUploadsLambda
+    timeout: 900 # 900 seconds = 15 minutes Lambda MAX
     memorySize: 2048
     ephemeralStorageSize: 1024
     maximumRetryAttempts: 0
