+++ services/uploads/src/avScan.test.ts
@@ -1,0 +1,17 @@
+describe('avScan', () => {
+    it('can be called', async () => {
+
+        // upload a file to s3
+
+
+        // invoke.
+
+        // check tags
+
+        throw new Error('NOPE')
+    })
+
+    it.todo('marks skipped for too big a file (config a smaller max size)')
+    it.todo('marks infected for an infected file')
+    it.todo('marks error if ClamAV errors')
+})
