--- services/uploads/src/deps/clamAV/index.ts
@@ -1,1 +1,0 @@
-export { NewClamAV, ClamAV } from './clamAV'
