--- services/uploads/src/deps/clamAV/testData/badDummy.pdf
@@ -1,1 +1,0 @@
-X5O!P%@AP[4\PZX54(P^)7CC)7}$EICAR-STANDARD-ANTIVIRUS-TEST-FILE!$H+H*
