--- services/uploads/src/deps/clamAV/testData/goodList.csv
@@ -1,1 +1,0 @@
-a,b,c,d
