--- services/uploads/src/lambdas/avScan.ts
@@ -1,58 +1,0 @@
-import { Context, S3Event } from 'aws-lambda'
-import { NewS3UploadsClient } from '../deps/s3'
-import { NewClamAV } from '../deps/clamAV'
-import { scanFile } from '../lib/avScan'
-
-async function avScan(event: S3Event, _context: Context) {
-    console.info('-----Start Antivirus Lambda function-----')
-
-    // Check on the values for our required config
-    const clamAVBucketName = process.env.CLAMAV_BUCKET_NAME
-    if (!clamAVBucketName || clamAVBucketName === '') {
-        throw new Error('Configuration Error: CLAMAV_BUCKET_NAME must be set')
-    }
-
-    const clamAVDefintionsPath = process.env.PATH_TO_AV_DEFINITIONS
-    if (!clamAVDefintionsPath || clamAVDefintionsPath === '') {
-        throw new Error(
-            'Configuration Error: PATH_TO_AV_DEFINITIONS must be set'
-        )
-    }
-
-    const maxFileSize = parseInt(process.env.MAX_FILE_SIZE || '314572800')
-
-    const s3Client = NewS3UploadsClient()
-
-    const clamAV = NewClamAV(
-        {
-            bucketName: clamAVBucketName,
-            definitionsPath: clamAVDefintionsPath,
-        },
-        s3Client
-    )
-
-    const record = event.Records[0]
-    if (!record) {
-        throw new Error('no record in request')
-    }
-
-    const s3ObjectKey = record.s3.object.key
-    const s3ObjectBucket = record.s3.bucket.name
-
-    console.info('Scanning ', s3ObjectKey, s3ObjectBucket)
-    const err = await scanFile(
-        s3Client,
-        clamAV,
-        s3ObjectKey,
-        s3ObjectBucket,
-        maxFileSize,
-        '/tmp/downloads'
-    )
-    if (err) {
-        throw err
-    }
-
-    return 'FILE SCANNED'
-}
-
-export { avScan }
