--- services/uploads/src/lambdas/index.ts
@@ -1,4 +1,0 @@
-export { avScan } from './avScan'
-export { avDownloadDefinitions } from './avDownloadDefinitions'
-export { avAuditUploads } from './avAuditUploads'
-export { avAuditFiles } from './avAuditFiles'
