--- services/uploads/src/testData/freshclam.conf
@@ -1,2 +1,0 @@
-DatabaseMirror database.clamav.net
-CompressLocalDatabase yes
