+++ test.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+set -e
+
+pushd tests
+./test.sh
+popd
